fix 6 done

+++ src/main/java/com/hhu20/jsgl/mapper/AdvisorsMapper.java
@@ -8,6 +8,6 @@ import java.util.Map;
 public interface AdvisorsMapper {
 
     List<String> selectTnoByTeamno(@Param("teamno") Integer teamno); //精确查询teamno 把该团队的指导老师的tno返回
-    void update(@Param("tcoefficient") float tcoefficient,@Param("tno") String tno,@Param("teamno") Integer teamno);
+    void update(@Param("tcoefficient") Double tcoefficient,@Param("tno") String tno,@Param("teamno") Integer teamno);
     List<Map> selectByTeamno(@Param("teamno") Integer teamno);
 }
